requirements.txt
albumentations>=1.3.0
numpy>=1.24.2
omegaconf>=2.3.0
pandas>=1.5.3
Pillow>=9.5.0
pillow_avif_plugin>=1.3.1
pytorch_lightning>=2.0.1.post0
PyYAML>=6.0
scikit_learn>=1.2.2
torch>=2.0.0
torchvision>=0.15.1
